ux_sxfx/backend/requirements.txt
Django>=5.0
djangorestframework
mysqlclient
django-cors-headers